backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
yt-dlp>=2024.12.0
httpx==0.28.1
slowapi==0.1.9
pydantic==2.10.0
python-dotenv==1.0.1
beautifulsoup4==4.12.3
pytest==8.3.4
pytest-asyncio==0.25.0
